Move import funcs to utils
